Update Selenium.pm
Fixed bug in get_string_array where instead of correctly returning an empty array when there are no results it returns an array containing a single empty string. The new code uses a negative lookbehind assertion to split on unescaped commas. It then uses a map to remove the escape characters from the list of strings.